Crud to car finished with role restriction

+++ src/main/java/com/security/repoSecurity/Service/Implementations/ImpCarroService.java
@@ -1,10 +1,12 @@
 package com.security.repoSecurity.Service.Implementations;
 
-import com.security.repoSecurity.Models.AutoMotor;
+import com.security.repoSecurity.DTO.CarroDTO;
 import com.security.repoSecurity.Models.Carro;
 import com.security.repoSecurity.Repository.CarroRepository;
 import com.security.repoSecurity.Service.Interface.ICarroService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,31 +18,39 @@ public class ImpCarroService implements ICarroService {
     private CarroRepository carroRepository;
 
     @Override
-    public void saveCarro(AutoMotor auto) {
-        Carro car = new Carro();
+    public void saveCarro(CarroDTO carroDTO) {
+      Carro carro = new Carro();
+      carro.setMarca(carroDTO.getMarca());
+      carro.setModelo(carroDTO.getModelo());
+      carro.setNumeroChazis(carroDTO.getNumeroChazis());
+      carro.setTipoCombustible(carroDTO.getTipoCombustible());
+      carro.setPrecio(carroDTO.getPrecio());
+      carro.setColor(carroDTO.getColor());
+      carro.setMillasRecorridas(carroDTO.getMillasRecorridas());
+      carro.setNumeroPuertas(carroDTO.getNumeroPuertas());
 
-        car.setMarca(auto.getMarca());
-        car.setModelo(auto.getModelo());
-        car.setNumeroChazis(auto.getNumeroChazis());
-        car.setTipoCombustible(auto.getTipoCombustible());
-        car.setPrecio(auto.getPrecio());
-        car.setColor(auto.getColor());
-        car.setMillasRecorridas(auto.getMillasRecorridas());
-        car.setNumeroPuertas(3);
-
-        carroRepository.save(car);
+      carroRepository.save(carro);
     }
 
     @Override
-    public String updateCarro(int id) {
-        /*Boolean encontrado = carroRepository.existsById(id);
-        if(encontrado == false){
-            return "este carro no existe";
-        }else {
-            //carroRepository.save()
-           carroRepository.findById(id);
-        }*/
-        return "hola";
+    public ResponseEntity<Carro> updateCarro(int id, CarroDTO carroDTO) {
+        Optional<Carro> CrudData = carroRepository.findById(id);
+        if (CrudData.isPresent()) {
+            Carro carro = CrudData.get();
+
+            carro.setMarca(carroDTO.getMarca());
+            carro.setModelo(carroDTO.getModelo());
+            carro.setNumeroChazis(carroDTO.getNumeroChazis());
+            carro.setTipoCombustible(carroDTO.getTipoCombustible());
+            carro.setPrecio(carroDTO.getPrecio());
+            carro.setColor(carroDTO.getColor());
+            carro.setMillasRecorridas(carroDTO.getMillasRecorridas());
+            carro.setNumeroPuertas(carroDTO.getNumeroPuertas());
+            return new ResponseEntity<>(carroRepository.save(carro), HttpStatus.OK);
+
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @Override
@@ -49,7 +59,14 @@ public class ImpCarroService implements ICarroService {
     }
 
     @Override
-    public void deleteCarro(int id) {
-        carroRepository.deleteById(id);
+    public ResponseEntity<?> deleteCarro(int id) {
+        Optional<Carro> CrudData = carroRepository.findById(id);
+
+        if(CrudData.isPresent()){
+            carroRepository.deleteById(id);
+            return new ResponseEntity<>("Carro con Id: " +id+ " eliminado", HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>("no se encontro el carro carro con Id: " +id+ " eliminado", HttpStatus.NOT_FOUND);
+        }
     }
 }
